Safari-609.1.11.1/LayoutTests/fast/events/onblur-remove-expected.txt

This tests that elements shouldn't emit any onblur events when they are being removed from the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS blurEventCount is 0

TEST COMPLETE
